requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
joblib==0.14.1
Markdown==3.1.1
numpy==1.18.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
sklearn==0.0
sqlparse==0.3.0
